requirements.txt
langchain==0.3.25  
langchain-core==0.3.59 
langchain-community==0.3.12 
langgraph==0.2.60 
# langsmith==0.2.5 
langchain-openai==0.2.12 
pydantic==2.9.2 
httpx==0.27.0
langchain-google-genai

fastapi==0.110.3
uvicorn[standard]==0.29.0
python-dotenv
jinja2>=3.1.0
